Move more tests from individual app view tests to the conversation view tests.
