+++ go/apps/bulk_message/tests/test_views.py
@@ -1,19 +1,10 @@
-from datetime import date
-from StringIO import StringIO
-from zipfile import ZipFile
-
-from django.core import mail
 from django.core.urlresolvers import reverse
-from django.utils.unittest import skip
 
 from vumi.tests.utils import RegexMatcher
 
 from go.vumitools.tests.utils import VumiApiCommand
 from go.vumitools.token_manager import TokenManager
 from go.apps.tests.base import DjangoGoApplicationTestCase
-from go.base.tests.utils import FakeMessageStoreClient, FakeMatchResult
-
-from mock import patch
 
 
 class BulkMessageTestCase(DjangoGoApplicationTestCase):
@@ -47,128 +38,6 @@ class BulkMessageTestCase(DjangoGoApplicationTestCase):
         self.assertEqual(conversation.name, self.TEST_CONVERSATION_NAME)
         self.assertContains(response, 'Write and send bulk message')
         self.assertContains(response, self.get_action_view_url('bulk_send'))
-
-    def test_show_cached_message_pagination(self):
-        self.setup_conversation()
-        # Create 21 inbound & 21 outbound messages, since we have
-        # 20 messages per page it should give us 2 pages
-        self.add_messages_to_conv(21)
-        response = self.client.get(self.get_view_url('message_list'))
-
-        # Check pagination
-        # Ordinarily we'd have 60 references to a contact, which by default
-        # display the from_addr if a contact cannot be found. (Each block has 3
-        # references, one in the table listing, 2 in the reply-to modal div.)
-        # We have no channels connected to this conversation, however, so we
-        # only have 20 in this test.
-        self.assertContains(response, 'from-', 20)
-        # We should have 2 links to page two, one for the actual page link
-        # and one for the 'Next' page link
-        self.assertContains(response, '&amp;p=2', 2)
-        # There should only be 1 link to the current page
-        self.assertContains(response, '&amp;p=1', 1)
-        # There should not be a link to the previous page since we are not
-        # the first page.
-        self.assertContains(response, '&amp;p=0', 0)
-
-    @skip("The new views don't have this.")
-    def test_show_cached_message_overview(self):
-        self.put_sample_messages_in_conversation(self.user_api,
-                                                 self.conv_key, 10)
-        response = self.client.get(self.get_view_url('show'))
-        self.assertContains(response,
-            '10 sent for delivery to the networks.')
-        self.assertContains(response,
-            '10 accepted for delivery by the networks.')
-        self.assertContains(response, '10 delivered.')
-
-    @skip("The new views don't have this.")
-    @patch('go.base.message_store_client.MatchResult')
-    @patch('go.base.message_store_client.Client')
-    def test_message_search(self, Client, MatchResult):
-        self.setup_conversation()
-        fake_client = FakeMessageStoreClient()
-        fake_result = FakeMatchResult()
-        Client.return_value = fake_client
-        MatchResult.return_value = fake_result
-
-        response = self.client.get(self.get_view_url('message_list'), {
-            'q': 'hello world 1',
-        })
-
-        template_names = [t.name for t in response.templates]
-        self.assertTrue(
-            'generic/includes/message-load-results.html' in template_names)
-        self.assertEqual(response.context['token'], fake_client.token)
-
-    @skip("The new views don't have this.")
-    @patch('go.base.message_store_client.MatchResult')
-    @patch('go.base.message_store_client.Client')
-    def test_message_results(self, Client, MatchResult):
-        self.setup_conversation()
-        fake_client = FakeMessageStoreClient()
-        fake_result = FakeMatchResult(tries=2,
-            results=[self.mkmsg_out() for i in range(10)])
-        Client.return_value = fake_client
-        MatchResult.return_value = fake_result
-
-        fetch_results_url = self.get_view_url('message_search_result')
-        fetch_results_params = {
-            'q': 'hello world 1',
-            'batch_id': 'batch-id',
-            'direction': 'inbound',
-            'token': fake_client.token,
-            'delay': 100,
-        }
-
-        response1 = self.client.get(fetch_results_url, fetch_results_params)
-        response2 = self.client.get(fetch_results_url, fetch_results_params)
-
-        # First time it should still show the loading page
-        self.assertTrue('generic/includes/message-load-results.html' in
-                        [t.name for t in response1.templates])
-        self.assertEqual(response1.context['delay'], 1.1 * 100)
-        # Second time it should still render the messages
-        self.assertTrue('generic/includes/message-list.html' in
-                        [t.name for t in response2.templates])
-        self.assertEqual(response1.context['token'], fake_client.token)
-        # Second time we should list the matching messages
-        self.assertEqual(response2.context['token'], fake_client.token)
-        self.assertEqual(
-            len(response2.context['message_page'].object_list), 10)
-
-    def test_aggregates(self):
-        self.setup_conversation(started=True)
-        self.add_messages_to_conv(
-            5, start_date=date(2012, 1, 1), time_multiplier=12)
-        response = self.client.get(self.get_view_url('aggregates'),
-                                   {'direction': 'inbound'})
-        self.assertEqual(response.content, '\r\n'.join([
-            '2011-12-30,1',
-            '2011-12-31,2',
-            '2012-01-01,2',
-            '',  # csv ends with a blank line
-            ]))
-
-    def test_export_messages(self):
-        self.setup_conversation(started=True)
-        self.add_messages_to_conv(
-            5, start_date=date(2012, 1, 1), time_multiplier=12, reply=True)
-        response = self.client.post(self.get_view_url('message_list'), {
-            '_export_conversation_messages': True,
-        })
-        self.assertRedirects(response, self.get_view_url('message_list'))
-        [email] = mail.outbox
-        self.assertEqual(email.recipients(), [self.django_user.email])
-        self.assertTrue(self.conversation.name in email.subject)
-        self.assertTrue(self.conversation.name in email.body)
-        [(file_name, zipcontent, mime_type)] = email.attachments
-        self.assertEqual(file_name, 'messages-export.zip')
-        zipfile = ZipFile(StringIO(zipcontent), 'r')
-        content = zipfile.open('messages-export.csv', 'r').read()
-        # 1 header, 5 sent, 5 received, 1 trailing newline == 12
-        self.assertEqual(12, len(content.split('\n')))
-        self.assertEqual(mime_type, 'application/zip')
 
     def test_action_bulk_send_view(self):
         self.setup_conversation(started=True, with_group=True,
@@ -271,11 +140,11 @@ class BulkMessageTestCase(DjangoGoApplicationTestCase):
                                 with_channel=True)
 
         # POST the action with a mock token manager
-        with patch.object(TokenManager, 'generate_token') as mock_method:
-            mock_method.return_value = ('abcdef', '123456')
-            response = self.client.post(
-                self.get_action_view_url('bulk_send'),
-                {'message': 'I am ham, not spam.', 'dedupe': True})
+        self.monkey_patch(
+            TokenManager, 'generate_token', lambda s: ('abcdef', '123456'))
+        response = self.client.post(
+            self.get_action_view_url('bulk_send'),
+            {'message': 'I am ham, not spam.', 'dedupe': True})
         self.assertRedirects(response, self.get_view_url('show'))
 
         # Check that we get a confirmation message
@@ -319,60 +188,3 @@ class BulkMessageTestCase(DjangoGoApplicationTestCase):
             batch_id=conversation.batch.key, msg_options={},
             delivery_class=conversation.delivery_class,
             content='I am ham, not spam.', dedupe=True))
-
-    @patch('go.vumitools.conversation.utils.ConversationWrapper.'
-           'has_channel_supporting_generic_sends')
-    def test_actions_on_inbound_only(self, hcsgs):
-        hcsgs.return_value = True
-        self.setup_conversation()
-        messages = self.add_messages_to_conv(1, reply=True)
-        [msg_in, msg_out] = messages[0]
-
-        response = self.client.get(
-            self.get_view_url('message_list'), {'direction': 'inbound'})
-        self.assertContains(response, 'Reply')
-        self.assertContains(response, 'href="#reply-%s"' % (
-            msg_in['message_id'],))
-
-        response = self.client.get(
-            self.get_view_url('message_list'), {'direction': 'outbound'})
-        self.assertNotContains(response, 'Reply')
-
-    def test_no_actions_on_inbound_with_no_generic_send_channels(self):
-        # We have no routing hooked up and hence no channels supporting generic
-        # sends.
-        self.setup_conversation()
-        messages = self.add_messages_to_conv(1, reply=True)
-        [msg_in, msg_out] = messages[0]
-
-        response = self.client.get(
-            self.get_view_url('message_list'), {'direction': 'inbound'})
-        self.assertNotContains(response, 'Reply')
-
-    def test_send_one_off_reply(self):
-        self.setup_conversation(started=True, with_group=True)
-        self.add_messages_to_conv(1)
-        conversation = self.get_wrapped_conv()
-        [msg] = conversation.received_messages()
-        response = self.client.post(self.get_view_url('message_list'), {
-            'in_reply_to': msg['message_id'],
-            'content': 'foo',
-            'to_addr': 'should be ignored',
-            '_send_one_off_reply': True,
-        })
-        self.assertRedirects(response, self.get_view_url('message_list'))
-
-        [reply_to_cmd] = self.get_api_commands_sent()
-        self.assertEqual(reply_to_cmd['worker_name'],
-                            'bulk_message_application')
-        self.assertEqual(reply_to_cmd['command'], 'send_message')
-        self.assertEqual(reply_to_cmd['args'],
-                         [self.conversation.user_account.key,
-                          self.conversation.key])
-        self.assertEqual(reply_to_cmd['kwargs']['command_data'], {
-            'batch_id': conversation.batch.key,
-            'conversation_key': conversation.key,
-            'content': 'foo',
-            'to_addr': msg['from_addr'],
-            'msg_options': {'in_reply_to': msg['message_id']},
-        })
