+++ go/apps/surveys/tests/test_views.py
@@ -1,16 +1,6 @@
-from datetime import date
-from zipfile import ZipFile
-from StringIO import StringIO
-
-from django.core import mail
-from django.utils.unittest import skip
-
 from go.vumitools.tests.utils import VumiApiCommand
 from go.apps.tests.base import DjangoGoApplicationTestCase
 from go.apps.surveys.view_definition import get_poll_config
-from go.base.tests.utils import FakeMessageStoreClient, FakeMatchResult
-
-from mock import patch
 
 
 class SurveyTestCase(DjangoGoApplicationTestCase):
@@ -180,117 +170,3 @@ class SurveyTestCase(DjangoGoApplicationTestCase):
         self.assertEqual(lines[0], 'user_id,user_timestamp,label-1,label-2')
         self.assertTrue(lines[1].startswith('user-1'))
         self.assertTrue(lines[1].endswith(',answer 1,answer 2'))
-
-    def test_aggregates(self):
-        self.setup_conversation(started=True)
-        self.add_messages_to_conv(
-            5, start_date=date(2012, 1, 1), time_multiplier=12)
-        response = self.client.get(self.get_view_url('aggregates'),
-                                   {'direction': 'inbound'})
-        self.assertEqual(response.content, '\r\n'.join([
-            '2011-12-30,1',
-            '2011-12-31,2',
-            '2012-01-01,2',
-            '',  # csv ends with a blank line
-        ]))
-
-    def test_export_messages(self):
-        self.setup_conversation(started=True)
-        self.add_messages_to_conv(
-            5, start_date=date(2012, 1, 1), time_multiplier=12, reply=True)
-        response = self.client.post(self.get_view_url('message_list'), {
-            '_export_conversation_messages': True,
-        })
-        self.assertRedirects(response, self.get_view_url('message_list'))
-        [email] = mail.outbox
-        self.assertEqual(email.recipients(), [self.django_user.email])
-        self.assertTrue(self.conversation.name in email.subject)
-        self.assertTrue(self.conversation.name in email.body)
-        [(file_name, contents, mime_type)] = email.attachments
-        self.assertEqual(file_name, 'messages-export.zip')
-
-        zipfile = ZipFile(StringIO(contents), 'r')
-        csv_contents = zipfile.open('messages-export.csv', 'r').read()
-
-        # 1 header, 5 sent, 5 received, 1 trailing newline == 12
-        self.assertEqual(12, len(csv_contents.split('\n')))
-        self.assertEqual(mime_type, 'application/zip')
-
-    @skip("The new views don't have this.")
-    @patch('go.base.message_store_client.MatchResult')
-    @patch('go.base.message_store_client.Client')
-    def test_message_search(self, Client, MatchResult):
-        fake_client = FakeMessageStoreClient()
-        fake_result = FakeMatchResult()
-        Client.return_value = fake_client
-        MatchResult.return_value = fake_result
-
-        response = self.client.get(self.get_view_url('message_list'), {
-            'q': 'hello world 1',
-        })
-
-        template_names = [t.name for t in response.templates]
-        self.assertTrue('generic/includes/message-load-results.html' in
-                        template_names)
-        self.assertEqual(response.context['token'], fake_client.token)
-
-    @skip("The new views don't have this.")
-    @patch('go.base.message_store_client.MatchResult')
-    @patch('go.base.message_store_client.Client')
-    def test_message_results(self, Client, MatchResult):
-        fake_client = FakeMessageStoreClient()
-        fake_result = FakeMatchResult(tries=2,
-            results=[self.mkmsg_out() for i in range(10)])
-        Client.return_value = fake_client
-        MatchResult.return_value = fake_result
-
-        fetch_results_params = {
-            'q': 'hello world 1',
-            'batch_id': 'batch-id',
-            'direction': 'inbound',
-            'token': fake_client.token,
-            'delay': 100,
-        }
-
-        response1 = self.client.get(self.get_view_url('message_search_result'),
-                                    fetch_results_params)
-        response2 = self.client.get(self.get_view_url('message_search_result'),
-                                    fetch_results_params)
-
-        # First time it should still show the loading page
-        self.assertTrue('generic/includes/message-load-results.html' in
-                            [t.name for t in response1.templates])
-        self.assertEqual(response1.context['delay'], 1.1 * 100)
-        # Second time it should still render the messages
-        self.assertTrue('generic/includes/message-list.html' in
-                            [t.name for t in response2.templates])
-        self.assertEqual(response1.context['token'], fake_client.token)
-        # Second time we should list the matching messages
-        self.assertEqual(response2.context['token'], fake_client.token)
-        self.assertEqual(len(response2.context['message_page'].object_list),
-            10)
-
-    def test_send_one_off_reply(self):
-        self.setup_conversation(started=True)
-        self.add_messages_to_conv(1)
-        conversation = self.get_wrapped_conv()
-        [msg] = conversation.received_messages()
-        response = self.client.post(self.get_view_url('message_list'), {
-            'in_reply_to': msg['message_id'],
-            'content': 'foo',
-            'to_addr': 'should be ignored',
-            '_send_one_off_reply': True,
-        })
-        self.assertRedirects(response, self.get_view_url('message_list'))
-
-        [reply_to_cmd] = self.get_api_commands_sent()
-        self.assertEqual(reply_to_cmd['worker_name'],
-                         'survey_application')
-        self.assertEqual(reply_to_cmd['command'], 'send_message')
-        self.assertEqual(reply_to_cmd['kwargs']['command_data'], {
-            'batch_id': conversation.batch.key,
-            'conversation_key': conversation.key,
-            'content': 'foo',
-            'to_addr': msg['from_addr'],
-            'msg_options': {'in_reply_to': msg['message_id']},
-        })
